fix/query_bugs
