+++ src/routes/graphql/mutationTypes/edit-user/edit-user.ts
@@ -1,7 +1,7 @@
 import { GraphQLObjectType } from "graphql";
-import {userObjectType} from "../../queryTypes/userQuery/userObjectType.js";
 import Context from "../../types/context.js";
 import { nonNullChangeUserObjectType, nonNullUUIDType } from "../../types/nonNullTypes.js";
+import {userObjectType} from "../../queryTypes/user-query/user-object-type.js";
 
 interface ChangeUser {
   id: string;
