+++ src/routes/graphql/queryTypes/post-query/post-object-type-list.ts
@@ -1,4 +1,4 @@
 import { GraphQLList } from 'graphql';
-import {postObjectType} from './postObjectType.js';
+import {postObjectType} from "./post-object-type.js";
 
 export const postObjectTypeList = new GraphQLList(postObjectType);
