--- src/routes/graphql/queryTypes/profileQuery/profileObjectTypeList.ts
@@ -1,5 +1,0 @@
-import { GraphQLList } from 'graphql';
-import {profileObjectType} from './profileObjectType.js';
-
-export const profileObjectTypeList = new GraphQLList(profileObjectType);
-
