--- src/routes/graphql/queryTypes/userQuery/userObjectTypeList.ts
@@ -1,5 +1,0 @@
-import { GraphQLList } from 'graphql';
-import {userObjectType} from './userObjectType.js';
-
-export const userObjectTypeList = new GraphQLList(userObjectType);
-
